renamed nti uc cached service and related code
