+++ ConcernsCaseWork/ConcernsCaseWork.Tests/Pages/Case/Management/Action/NtiUc/IndexPageModelTests.cs
@@ -8,7 +8,7 @@ using Microsoft.AspNetCore.Mvc.ViewFeatures;
 using Microsoft.Extensions.Logging;
 using Moq;
 using NUnit.Framework;
-using Service.Redis.Nti;
+using Service.Redis.NtiUnderConsideration;
 using System;
 using System.Threading.Tasks;
