+++ ConcernsCaseWork/ConcernsCaseWork.Tests/Pages/Case/Management/IndexPageModelTests.cs
@@ -13,7 +13,7 @@ using Microsoft.AspNetCore.Mvc.ViewFeatures;
 using Microsoft.Extensions.Logging;
 using Moq;
 using NUnit.Framework;
-using Service.Redis.Nti;
+using Service.Redis.NtiUnderConsideration;
 using Service.Redis.Status;
 using Service.TRAMS.Status;
 using System.Linq;
