+++ ConcernsCaseWork/ConcernsCaseWork/Services/Nti/NtiModelService.cs
@@ -1,6 +1,6 @@
 ﻿using ConcernsCaseWork.Mappers;
 using ConcernsCaseWork.Models.CaseActions;
-using Service.Redis.Nti;
+using Service.Redis.NtiUnderConsideration;
 using System;
 using System.Collections.Generic;
 using System.Linq;
