+++ ConcernsCaseWork/Service.Redis/NtiUnderConsideration/INtiStatusesCachedService.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Service.Redis.Nti
+namespace Service.Redis.NtiUnderConsideration
 {
 	public interface INtiStatusesCachedService
 	{
