+++ ConcernsCaseWork/Service.Redis/NtiUnderConsideration/NtiStatusesCachedService.cs
@@ -6,7 +6,7 @@ using System.Collections.Generic;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Service.Redis.Nti
+namespace Service.Redis.NtiUnderConsideration
 {
 	public class NtiStatusesCachedService : CachedService, INtiStatusesCachedService
 	{
